requirements.txt
blinker==1.8.2
click==8.1.7
dnspython==2.6.1
email-validator==2.1.1
Flask==3.0.3
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
itsdangerous==2.2.0
Jinja2==3.1.4
SQLAlchemy==2.0.30
Werkzeug==3.0.3
WTForms==3.1.2
python-dotenv==1.1.1